+++ app/Models/User.php
@@ -60,7 +60,7 @@ class User extends Authenticatable
      *
      * @var bool
      */
-    public $timestamps = false;
+    public $timestamps = true;
 
     /**
      * Returns the login identifier field name.